Splitting OSX and Linux packages installation flows.
'brew' and Linux package managers have a different implementation and
sometimes a different behavior, for example, we need to 'brew update'
before we 'brew upgrade' while in Linux 'dnf update' is deprecated and
'dnf upgrade' replaces it.

In addition, when it comes to error handling it is simpler to split the
flows.

Signed-off-by: Yoni Bettan <ybettan@redhat.com>