Explicitly using the fedora image from `fedoraproject.org`.
When running locally with podman/docker, `fedora:latest` will default to
`registry.fedoraproject.org/fedora:latest` but on Ubuntu it will default
to `docker.io/library/fedora:latest` which contain a different `dnf`
version it is will create inconsistencies.

With that change we can make sure we always use the same fedora image
for CI testing.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>